refactor(gateway): move attempt resilience into gateway_pipeline_attempt.go
